requirements.txt
crewai[tools]>=0.95.0
google-genai>=1.0.0
tavily-python>=0.5.0
python-dotenv>=1.0.0
markdown>=3.5.0
pdfkit>=1.0.0
litellm>=1.50.0
pydantic>=2.0.0
